17-feb 11am - Nuevo cambio de estructuración + refactoring reader class
