+++ pyplexos/Protocols.py
@@ -1,16 +1,20 @@
 from typing import Protocol
 from datetime import datetime
 
+
 class PlexosReaderProtocol(Protocol):
     @property
     def get_solution_model(self) -> dict[str, list[dict]]:
         pass
+
     @property
     def get_solution_data(self) -> dict[str, dict[str, list]]:
         pass
+
     @property
-    def get_min_datetime(self) -> datetime:
+    def get_initial_datetime(self) -> datetime:
         pass
+
 
 class WriterModelProtocol(Protocol):
     def write(self, data: PlexosReaderProtocol) -> None:
